remove old test and add new ones
