+++ test/pettingzoo/test_pistonball.py
@@ -1,9 +1,9 @@
 import pprint
 
-from tic_tac_toe import get_args, train_agent, watch
+from pistonball import get_args, train_agent, watch
 
 
-def test_tic_tac_toe(args=get_args()):
+def test_piston_ball(args=get_args()):
     if args.watch:
         watch(args)
         return
@@ -18,4 +18,4 @@ def test_tic_tac_toe(args=get_args()):
 
 
 if __name__ == '__main__':
-    test_tic_tac_toe(get_args())
+    test_piston_ball(get_args())
